Use single function to configure all three cloud proxies
