+++ crates/core/tedge_mapper/src/c8y/mapper.rs
@@ -1,5 +1,6 @@
 use crate::core::component::TEdgeComponent;
 use crate::core::mapper::start_basic_actors;
+use crate::core::mqtt::configure_proxy;
 use anyhow::Context;
 use async_trait::async_trait;
 use c8y_api::http_proxy::read_c8y_credentials;
@@ -12,21 +13,14 @@ use c8y_mapper_ext::config::C8yMapperConfig;
 use c8y_mapper_ext::converter::CumulocityConverter;
 use mqtt_channel::Config;
 use std::borrow::Cow;
-use std::sync::Arc;
 use tedge_api::entity::EntityExternalId;
 use tedge_api::mqtt_topics::EntityTopicId;
-use tedge_config::all_or_nothing;
-use tedge_config::models::proxy_scheme::ProxyScheme;
 use tedge_config::tedge_toml::ProfileName;
 use tedge_config::TEdgeConfig;
 use tedge_downloader_ext::DownloaderActor;
 use tedge_file_system_ext::FsWatchActorBuilder;
 use tedge_http_ext::HttpActor;
 use tedge_mqtt_bridge::rumqttc::LastWill;
-use tedge_mqtt_bridge::rumqttc::Proxy;
-use tedge_mqtt_bridge::rumqttc::ProxyAuth;
-use tedge_mqtt_bridge::rumqttc::ProxyType;
-use tedge_mqtt_bridge::rumqttc::TlsConfiguration;
 use tedge_mqtt_bridge::rumqttc::Transport;
 use tedge_mqtt_bridge::use_credentials;
 use tedge_mqtt_bridge::BridgeConfig;
@@ -228,27 +222,7 @@ impl TEdgeComponent for CumulocityMapper {
             });
             cloud_config.set_keep_alive(c8y_config.bridge.keepalive_interval.duration());
 
-            let rustls_config = tedge_config.cloud_client_tls_config();
-            let proxy_config = &tedge_config.proxy;
-            if let Some(address) = proxy_config.address.or_none() {
-                let credentials =
-                    all_or_nothing((proxy_config.username.clone(), proxy_config.password.clone()))
-                        .map_err(|e| anyhow::anyhow!(e))?;
-                cloud_config.set_proxy(Proxy {
-                    addr: address.host().to_string(),
-                    port: address.port().0,
-                    auth: match credentials {
-                        Some((username, password)) => ProxyAuth::Basic { username, password },
-                        None => ProxyAuth::None,
-                    },
-                    ty: match address.scheme() {
-                        ProxyScheme::Http => ProxyType::Http,
-                        ProxyScheme::Https => {
-                            ProxyType::Https(TlsConfiguration::Rustls(Arc::new(rustls_config)))
-                        }
-                    },
-                });
-            }
+            configure_proxy(&tedge_config, &mut cloud_config)?;
 
             runtime
                 .spawn(
