+++ crates/core/tedge_mapper/src/core/mod.rs
@@ -1,2 +1,3 @@
 pub mod component;
 pub mod mapper;
+pub mod mqtt;
